+++ Tests/SentryTests/SentrySpanTests.swift
@@ -74,7 +74,9 @@ class SentrySpanTests: XCTestCase {
         let span = fixture.getSut(client: client)
         let childSpan = span.startChild(operation: fixture.someOperation)
         
+        childSpan.finish()
         span.finish()
+        
         let lastEvent = client.captureEventWithScopeArguments[0].event
         let serializedData = lastEvent.serialize()
         
@@ -83,6 +85,20 @@ class SentrySpanTests: XCTestCase {
         
         XCTAssertEqual(serializedChild["span_id"] as? String, childSpan.context.spanId.sentrySpanIdString)
         XCTAssertEqual(serializedChild["parent_span_id"] as? String, span.context.spanId.sentrySpanIdString)
+    }
+    
+    func testFinishWithUnfinishedChild() {
+        let client = TestClient(options: fixture.options)!
+        let span = fixture.getSut(client: client)
+        span.startChild(operation: fixture.someOperation)
+        
+        span.finish()
+        let lastEvent = client.captureEventWithScopeArguments[0].event
+        let serializedData = lastEvent.serialize()
+        
+        let spans = serializedData["spans"] as! [Any]
+        
+        XCTAssertEqual(spans.count, 0)
     }
     
     func testStartChildWithNameOperation() {